+++ src/tests/audit_as_crates_io.rs
@@ -58,6 +58,17 @@ fn get_audit_as_crates_io(cfg: &Config, store: &Store, add_packages_to_index: bo
     }
 }
 
+fn get_audit_as_crates_io_no_network(cfg: &Config, store: &Store) -> String {
+    let mut cache = crate::storage::Cache::acquire(cfg).unwrap();
+    let res = tokio::runtime::Handle::current().block_on(crate::check_audit_as_crates_io(
+        cfg, store, None, &mut cache,
+    ));
+    match res {
+        Ok(()) => String::new(),
+        Err(e) => format!("{:?}", miette::Report::new(e)),
+    }
+}
+
 fn get_audit_as_crates_io_json(cfg: &Config, store: &Store) -> String {
     let mut cache = crate::storage::Cache::acquire(cfg).unwrap();
     let mut network = crate::network::Network::new_mock();
@@ -91,6 +102,47 @@ fn simple_audit_as_crates_io() {
 
     let output = get_audit_as_crates_io(&cfg, &store, false);
     insta::assert_snapshot!("simple-audit-as-crates-io", output);
+}
+
+#[test]
+fn simple_audit_as_crates_io_no_network() {
+    let _enter = TEST_RUNTIME.enter();
+
+    let mock = MockMetadata::simple();
+    let metadata = mock.metadata();
+    let (config, audits, imports) = builtin_files_full_audited(&metadata);
+    let store = Store::mock(config, audits, imports);
+    let cfg = mock_cfg(&metadata);
+
+    let output = get_audit_as_crates_io_no_network(&cfg, &store);
+    insta::assert_snapshot!("simple-audit-as-crates-io-no-network", output);
+}
+
+#[test]
+fn simple_audit_as_crates_io_no_network_existing_audits() {
+    let _enter = TEST_RUNTIME.enter();
+
+    let mock = MockMetadata::simple();
+    let metadata = mock.metadata();
+    let (config, mut audits, imports) = builtin_files_full_audited(&metadata);
+    audits
+        .audits
+        .entry("first-party".to_owned())
+        .or_default()
+        .push(full_audit(ver(10), SAFE_TO_DEPLOY));
+    {
+        let root_pkg_entry = audits.audits.entry("root-package".to_owned()).or_default();
+        root_pkg_entry.push(full_audit(ver(9), SAFE_TO_DEPLOY));
+        root_pkg_entry.push(delta_audit(ver(9), ver(10), SAFE_TO_DEPLOY));
+    }
+    let store = Store::mock(config, audits, imports);
+    let cfg = mock_cfg(&metadata);
+
+    let output = get_audit_as_crates_io_no_network(&cfg, &store);
+    insta::assert_snapshot!(
+        "simple-audit-as-crates-io-no-network-existing-audit",
+        output
+    );
 }
 
 #[test]