+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
-Nothing yet
+* `cargo vet check --frozen` (a disabled network) will infer the existence of crates using versions
+  in audits when checking audit-as-crates-io policies (#661)
 
 # Version 0.10.1 (2025-02-10)